Resolve 'Key not found: lastModifiedDateTime' (OneDrive API change)
* Original patch (90505dc800b48db1d17deb27d16b15c50a5b09ac) whilst it
resolved the issue of a OneDrive API change, when a file was deleted on
OneDrive, it was not actually deleted on the local file system. This
patch re-implements the fix for 'Key not found: lastModifiedDateTime' by
taking into account the deleted state of the item and correctly removes
the local file that is now in the deleted state.